// src/views/FarmView/commons/FarmAshSwap/type.graphql.ts
export class Farm {
  __typename?: "Farm";
  address: string;
  farmToken: Token;
  rewardToken: Token;
  farmingToken: Token;
  farmTokenSupply?: Nullable<string>;
  state: string;
  rewardPerSec?: Nullable<string>;
  rewardPerShare?: Nullable<string>;
  lastRewardBlockTs?: Nullable<number>;
  divisionSafetyConstant?: Nullable<string>;
  farmingTokenBalance?: Nullable<string>;
  produceRewardEnabled?: Nullable<boolean>;
  shard?: Nullable<string>;
  additionalRewards: AdditionalReward[];
  account?: Nullable<FarmAccount>;
}

export class AdditionalReward {
  __typename?: "AdditionalReward";
  rewardPerSec: string;
  rewardPerShare: string;
  periodRewardEnd: number;
  tokenId: string;
}

export class FarmAccount {
  __typename?: "FarmAccount";
  slopeBoosted?: Nullable<string>;
}
// ...
export class Pool {
  __typename?: "Pool";
  address: string;
  lpToken: Token;
  tokens: Token[];
  reserves: string[];
  underlyingPrices: string[];
  totalSupply: string;
  swapFeePercent: string;
  adminFeePercent: string;
  ampFactor: string;
  state: string;
  statistic?: Nullable<PoolStatistic>;
  graphStatistic?: Nullable<Nullable<Nullable<number>[]>[]>;
  apr?: Nullable<PoolApr>;
}
// ...
export class PoolStatistic {
  __typename?: "PoolStatistic";
  token_2_admin_fee_usd?: Nullable<number>;
  timestamp?: Nullable<number>;
  token_3_admin_fee_usd?: Nullable<number>;
  volume_usd?: Nullable<number>;
  token_2_amount?: Nullable<number>;
  tvl?: Nullable<number>;
  token_3_total_fee_usd?: Nullable<number>;
  address?: Nullable<string>;
  token_3_amount_usd?: Nullable<number>;
  token_1_admin_fee_usd?: Nullable<number>;
  token_1_amount_usd?: Nullable<number>;
  token_2_total_fee_usd?: Nullable<number>;
  unique_traders?: Nullable<number>;
  token_1_total_fee_usd?: Nullable<number>;
  token_3_amount?: Nullable<number>;
  apr?: Nullable<number>;
  transaction_count?: Nullable<number>;
  token_1_amount?: Nullable<number>;
  token_2_amount_usd?: Nullable<number>;
}

export class LiquidityDistribution {
  __typename?: "LiquidityDistribution";
  liquidity?: Nullable<number>;
  token?: Nullable<string>;
}

export class PoolApr {
  __typename?: "PoolApr";
  address?: Nullable<string>;
  apr?: Nullable<Nullable<Nullable<number>[]>[]>;
}
// ...
export class Token {
  __typename?: "Token";
  id?: Nullable<string>;
  price?: Nullable<number>;
  statistic?: Nullable<TokenStatistic>;
  graphStatistic?: Nullable<Nullable<Nullable<number>[]>[]>;
  tokenFee?: Nullable<TokenFee>;
  tokenPool?: Nullable<Nullable<TokenPool>[]>;
}

export class TokenStatistic {
  __typename?: "TokenStatistic";
  change_percentage_day: number;
  change_percentage_hour: number;
  change_percentage_week: number;
  liquidity: number;
  price: number;
  transaction_count: number;
  volume: number;
}

export class TokenFee {
  __typename?: "TokenFee";
  fee?: Nullable<Nullable<Nullable<number>[]>[]>;
  token?: Nullable<string>;
}
// ...
export class TokenPool {
  __typename?: "TokenPool";
  address?: Nullable<string>;
  apr?: Nullable<number>;
  timestamp?: Nullable<string>;
  token_1_admin_fee_usd?: Nullable<number>;
  token_1_amount?: Nullable<string>;
  token_1_amount_usd?: Nullable<number>;
  token_1_total_fee_usd?: Nullable<number>;
  token_2_admin_fee_usd?: Nullable<number>;
  token_2_amount?: Nullable<string>;
  token_2_amount_usd?: Nullable<number>;
  token_2_total_fee_usd?: Nullable<number>;
  token_3_admin_fee_usd?: Nullable<string>;
  token_3_amount?: Nullable<string>;
  token_3_amount_usd?: Nullable<number>;
  token_3_total_fee_usd?: Nullable<number>;
  transaction_count?: Nullable<number>;
  tvl?: Nullable<number>;
  unique_traders?: Nullable<string>;
  volume_usd?: Nullable<number>;
}

export class VotingEscrow {
  __typename?: "VotingEscrow";
  address: string;
  lockedToken: Token;
  totalLock?: Nullable<string>;
  veSupply?: Nullable<string>;
  account?: Nullable<VeAccount>;
}

export class Locked {
  __typename?: "Locked";
  amount?: Nullable<string>;
  end?: Nullable<string>;
}

export class VeAccount {
  __typename?: "VeAccount";
  locked?: Nullable<Locked>;
}

type Nullable<T> = T | null;